requirements.txt
huggingface_hub
jaxtyping
matplotlib
numpy
python-dotenv
safetensors
seaborn
sklearn
torch
tqdm
transformers
transformer_lens
typing